Rather far-reaching changes in the way acoustic scores are handled. Since senone scores are quantized, we might as well keep the acoustic scores quantized as long as possible.  They won't quite fit in 8 bits (particularly for 1s_c_d_dd features) but 16 bits are more than enough.
Also totally rework the multi-stream Gaussian computation so it
(hopefully) works in fixed-point mode as well.

Currently, the HMM evaluation code is still using 32-bit Viterbi
scores but there's probably no good reason for that.

Quantized codebooks are coming at some point soon, though it's not
clear they'll help semi-continuous models much, and we need to
implement Gaussian selection for continuous ones first probably.


git-svn-id: svn+ssh://svn.code.sf.net/p/cmusphinx/code/trunk/pocketsphinx@7879 94700074-3cef-4d97-a70e-9c8c206c02f5

+++ src/libpocketsphinx/ms_mgau.h
@@ -128,7 +128,7 @@ typedef struct {
 
     /**< Intermediate used in computation */
     gauden_dist_t ***dist;  
-    int8 *mgau_active;
+    uint8 *mgau_active;
 
 } ms_mgau_model_t;  
 
@@ -144,7 +144,7 @@ void ms_mgau_free(ms_mgau_model_t *g /**< In: A set of models to free */
     );
 
 int32 ms_cont_mgau_frame_eval(ms_mgau_model_t * msg,
-                              int32 *senscr,
+                              int16 *senscr,
                               int32 *senone_active,
                               int32 n_senone_active,
                               mfcc_t ** feat,
